+++ python_modules/dagster/dagster/core/execution.py
@@ -37,12 +37,17 @@ from .execution_context import (
     SystemPipelineExecutionContext,
 )
 
-
 from .errors import DagsterInvariantViolationError
 
 from .events import construct_event_logger
 
 from .execution_plan.create import create_execution_plan_core
+
+from .execution_plan.intermediates_manager import (
+    FileSystemIntermediatesManager,
+    InMemoryIntermediatesManager,
+    IntermediatesManager,
+)
 
 from .execution_plan.objects import (
     ExecutionPlan,
@@ -55,13 +60,17 @@ from .execution_plan.multiprocessing_engine import multiprocess_execute_plan
 
 from .execution_plan.simple_engine import start_inprocess_executor
 
-from .files import LocalTempFileStore
-
 from .init_context import InitContext, InitResourceContext
 
 from .log import DagsterLog
 
-from .runs import DagsterRunMeta
+from .runs import (
+    DagsterRunMeta,
+    FileSystemRunStorage,
+    InMemoryRunStorage,
+    RunStorage,
+    RunStorageMode,
+)
 
 from .system_config.objects import EnvironmentConfig
 
@@ -346,29 +355,90 @@ def _create_persistence_strategy(persistence_config):
         check.failed('Unsupported persistence key: {}'.format(persistence_key))
 
 
+def construct_run_storage(run_config, environment_config):
+    '''
+    Construct the run storage for this pipeline. Our rules are the following:
+
+    If the called has specified a run_storage_factory_fn, we use for creating
+    the run storage.
+
+    Then we fallback to config.
+
+    If there is no config, we default to in memory storage. This is mostly so
+    that tests default to in-memory.
+    '''
+    check.inst_param(run_config, 'run_config', RunConfig)
+    check.inst_param(environment_config, 'environment_config', EnvironmentConfig)
+
+    if run_config.storage_mode:
+        if run_config.storage_mode == RunStorageMode.FILESYSTEM:
+            return FileSystemRunStorage()
+        elif run_config.storage_mode == RunStorageMode.IN_MEMORY:
+            return InMemoryRunStorage()
+        else:
+            check.failed('Unexpected enum {}'.format(run_config.storage_mode))
+    elif environment_config.storage.storage_mode == 'filesystem':
+        return FileSystemRunStorage()
+    elif environment_config.storage.storage_mode == 'in_memory':
+        return InMemoryRunStorage()
+    elif environment_config.storage.storage_mode is None:
+        return InMemoryRunStorage()
+    else:
+        raise DagsterInvariantViolationError(
+            'Invalid storage specified {}'.format(environment_config.storage.storage_mode)
+        )
+
+
+def construct_intermediates_manager(run_config, init_context, environment_config):
+    check.inst_param(run_config, 'run_config', RunConfig)
+    check.inst_param(init_context, 'init_context', InitContext)
+    check.inst_param(environment_config, 'environment_config', EnvironmentConfig)
+
+    if run_config.storage_mode:
+        if run_config.storage_mode == RunStorageMode.FILESYSTEM:
+            return FileSystemIntermediatesManager(init_context.run_id)
+        elif run_config.storage_mode == RunStorageMode.IN_MEMORY:
+            return InMemoryIntermediatesManager()
+        else:
+            check.failed('Unexpected enum {}'.format(run_config.storage_mode))
+    elif environment_config.storage.storage_mode == 'filesystem':
+
+        return FileSystemIntermediatesManager(init_context.run_id)
+    elif environment_config.storage.storage_mode == 'in_memory':
+        return InMemoryIntermediatesManager()
+    elif environment_config.storage.storage_mode is None:
+        return InMemoryIntermediatesManager()
+    else:
+        raise DagsterInvariantViolationError(
+            'Invalid storage specified {}'.format(environment_config.storage.storage_mode)
+        )
+
+
 @contextmanager
 def yield_pipeline_execution_context(pipeline_def, environment_dict, run_config):
     check.inst_param(pipeline_def, 'pipeline_def', PipelineDefinition)
     check.dict_param(environment_dict, 'environment_dict', key_type=str)
     check.inst_param(run_config, 'run_config', RunConfig)
 
-    run_config.run_storage.write_dagster_run_meta(
+    environment_config = create_environment_config(pipeline_def, environment_dict)
+
+    context_definition = pipeline_def.context_definitions[environment_config.context.name]
+
+    run_storage = construct_run_storage(run_config, environment_config)
+
+    run_storage.write_dagster_run_meta(
         DagsterRunMeta(
             run_id=run_config.run_id, timestamp=time.time(), pipeline_name=pipeline_def.name
         )
     )
 
-    environment_config = create_environment_config(pipeline_def, environment_dict)
-
-    context_definition = pipeline_def.context_definitions[environment_config.context.name]
-
-    ec_or_gen = context_definition.context_fn(
-        InitContext(
-            context_config=environment_config.context.config,
-            pipeline_def=pipeline_def,
-            run_id=run_config.run_id,
-        )
+    init_context = InitContext(
+        context_config=environment_config.context.config,
+        pipeline_def=pipeline_def,
+        run_id=run_config.run_id,
     )
+
+    ec_or_gen = context_definition.context_fn(init_context)
 
     with as_ensured_single_gen(ec_or_gen) as execution_context:
         check.inst(execution_context, ExecutionContext)
@@ -381,17 +451,31 @@ def yield_pipeline_execution_context(pipeline_def, environment_dict, run_config)
             run_config.run_id,
         ) as resources:
             yield construct_pipeline_execution_context(
-                run_config, execution_context, pipeline_def, resources, environment_config
+                run_config,
+                execution_context,
+                pipeline_def,
+                resources,
+                environment_config,
+                run_storage,
+                construct_intermediates_manager(run_config, init_context, environment_config),
             )
 
 
 def construct_pipeline_execution_context(
-    run_config, execution_context, pipeline, resources, environment_config
+    run_config,
+    execution_context,
+    pipeline,
+    resources,
+    environment_config,
+    run_storage,
+    intermediates_manager,
 ):
     check.inst_param(run_config, 'run_config', RunConfig)
     check.inst_param(execution_context, 'execution_context', ExecutionContext)
     check.inst_param(pipeline, 'pipeline', PipelineDefinition)
     check.inst_param(environment_config, 'environment_config', EnvironmentConfig)
+    check.inst_param(run_storage, 'run_storage', RunStorage)
+    check.inst_param(intermediates_manager, 'intermediates_manager', IntermediatesManager)
 
     loggers = _create_loggers(run_config, execution_context)
     tags = get_tags(execution_context, run_config, pipeline)
@@ -406,7 +490,8 @@ def construct_pipeline_execution_context(
             persistence_strategy=_create_persistence_strategy(
                 environment_config.context.persistence
             ),
-            files=LocalTempFileStore(run_config.run_id),
+            run_storage=run_storage,
+            intermediates_manager=intermediates_manager,
         ),
         tags=tags,
         log=log,
@@ -589,7 +674,7 @@ def invoke_executor_on_plan(pipeline_context, execution_plan, step_keys_to_execu
         step_events_gen = start_inprocess_executor(
             pipeline_context,
             execution_plan,
-            pipeline_context.run_config.executor_config.inmem_intermediates_manager,
+            pipeline_context.intermediates_manager,
             step_keys_to_execute,
         )
     elif isinstance(pipeline_context.executor_config, MultiprocessExecutorConfig):