+++ python_modules/dagster/dagster/core/execution_plan/intermediates_manager.py
@@ -1,12 +1,14 @@
 from abc import ABCMeta, abstractmethod
 from collections import namedtuple
+from contextlib import contextmanager
 import pickle
+import os
 
 import six
 
 from dagster import check
-
-from dagster.core.files import FileStore
+from dagster import seven
+from dagster.utils import mkdir_p
 
 
 class StepOutputHandle(namedtuple('_StepOutputHandle', 'step_key output_name')):
@@ -78,15 +80,9 @@ class InMemoryIntermediatesManager(IntermediatesManager):
         return step_output_handle in self.values
 
 
-# TODO: This should go through persistence and serialization infrastructure
-# Fixing this requires getting the type information (serialization_strategy is
-# a property of runtime type) of the step output you are dealing with. As things
-# are structured now we construct the inputs before execution. The fix to this
-# will likely be to inject a different type of execution step that threads
-# the manager
-class FileSystemIntermediateManager(IntermediatesManager):
-    def __init__(self, files):
-        self._files = check.inst_param(files, 'files', FileStore)
+class FileSystemIntermediatesManager(IntermediatesManager):
+    def __init__(self, run_id):
+        self._files = LocalTempFileStore(run_id)
 
     def _get_path_comps(self, step_output_handle):
         return ['intermediates', step_output_handle.step_key, step_output_handle.output_name]
@@ -107,3 +103,48 @@ class FileSystemIntermediateManager(IntermediatesManager):
 
     def has_value(self, step_output_handle):
         return self._files.has_file(*self._get_path_comps(step_output_handle))
+
+
+def check_path_comps(path_comps):
+    path_list = check.list_param(list(path_comps), 'path_comps', of_type=str)
+    check.param_invariant(path_list, 'path_list', 'Must have at least one comp')
+    return path_list
+
+
+class LocalTempFileStore:
+    def __init__(self, run_id):
+        check.str_param(run_id, 'run_id')
+        self.root = os.path.join(
+            seven.get_system_temp_directory(), 'dagster', 'runs', run_id, 'files'
+        )
+
+    @contextmanager
+    def writeable_binary_stream(self, *path_comps):
+        path_list = check_path_comps(path_comps)
+
+        target_dir = os.path.join(self.root, *path_list[:-1])
+        mkdir_p(target_dir)
+
+        target_path = os.path.join(target_dir, path_list[-1])
+        check.invariant(not os.path.exists(target_path))
+        with open(target_path, 'wb') as ff:
+            yield ff
+
+    @contextmanager
+    def readable_binary_stream(self, *path_comps):
+        path_list = check_path_comps(path_comps)
+
+        target_path = os.path.join(self.root, *path_list)
+        with open(target_path, 'rb') as ff:
+            yield ff
+
+    def has_file(self, *path_comps):
+        path_list = check_path_comps(path_comps)
+
+        target_path = os.path.join(self.root, *path_list)
+
+        if os.path.exists(target_path):
+            check.invariant(os.path.isfile(target_path))
+            return True
+        else:
+            return False